app/requirements.txt
streamlit>=1.11.1
ladybug-vtk>=0.8.1
ladybug-core
dragonfly-core
pollination-streamlit>=0.5.0
pollination-streamlit-io>=0.36.6
pollination-streamlit-viewer>=0.4.4
honeybee-vtk>=0.38.3
lbt-honeybee>=0.5.482